cleanup; remove old classes; fix messages; update README.md
