+++ Transit.cs
@@ -1,154 +1,95 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Net;
-using System.Net.Sockets;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using System.Net;
+using System.IO;
+using System.Threading;
+using static WebOne.Program;
+using System.Text.RegularExpressions;
 
 namespace WebOne
 {
-	/// <summary>
-	/// Транзитная передача HTTP-контента с исправлением содержимого
-	/// </summary>
+/// <summary>
+/// Transit transfer of content with applying some edits
+/// </summary>
 	class Transit
 	{
-		static string LastURL = "http://999.999.999.999/CON";//dirty workaround for phantom.sannata.org and similar sites
+		HttpListenerRequest ClientRequest;
+		HttpListenerResponse ClientResponse;
+		DateTime BeginTime;
+
+
 		byte[] UTF8BOM = Encoding.UTF8.GetPreamble();
 
-		byte[] Buffer = new byte[ConfigFile.RequestBufferSize]; //todo: adjust for real request size
-		string Request = "";
-		string RequestHeaders = "";
-		string RequestBody = "";
-		int RequestHeadersEnd;
-		string RequestUri = "about:blank";
-		bool LocalMode = false;
+		Uri RequestURL = new Uri("about:blank");
+		static string LastURL = "http://999.999.999.999/CON";
 		bool ShouldRedirectInNETFW = false;
 
 		HttpResponse response;
 		int ResponseCode = 502;
-		string ResponseHeaders;
 		string ResponseBody = ":(";
-		byte[] ResponseBuffer;
 		Stream TransitStream = null;
-		string ContentType = "";
+		string ContentType = "text/plain";
 
-		//Based on https://habr.com/ru/post/120157/
-		//Probably the class name needs to be changed
 
 		/// <summary>
 		/// Convert a Web 2.0 page to Web 1.0-like page.
 		/// </summary>
-		/// <param name="Client">TcpListener client</param>
-		public Transit(TcpClient Client)
+		/// <param name="ClientRequest">Request from HttpListener</param>
+		/// <param name="ClientResponse">Response for HttpListener</param>
+		/// <param name="BeginTime">Initial time (for log)</param>
+		public Transit(HttpListenerRequest ClientRequest, HttpListenerResponse ClientResponse, DateTime BeginTime)
 		{
-			ShouldRedirectInNETFW = false;
-			Console.Write("\n>");
+			this.ClientRequest = ClientRequest;
+			this.ClientResponse = ClientResponse;
+			this.BeginTime = BeginTime;
+			#if DEBUG
+			Console.WriteLine("{0}\t Begin process.", GetTime(BeginTime));
+			#endif
 			try
 			{
 
-				int Count = 0;
-				NetworkStream ClientStream;
-				try
-				{
-					//get request
-					ClientStream = Client.GetStream();
-
-					for (int i = 0; i < ConfigFile.ClientTimeout; i++) { Console.CursorLeft = Console.CursorLeft; if (ClientStream.DataAvailable) break; } //wait for slow clients
-																																						   //while (true) { if (ClientStream.DataAvailable) break; }
-
-					while (ClientStream.DataAvailable)
-					{
-						ClientStream.Read(Buffer, Count, 1);
-						Count++;
-					}
-					Request += Encoding.Default.GetString(Buffer).Trim('\0'); //cut empty 10 megabytes
-					//todo: внедрить адаптируемый к размеру запроса вариант: http://www.cyberforum.ru/post6920071.html
-					//or replacу with HttpListener (or something).
-
-					//check if the request is empty and report to try again.
-					if (Request.Length == 0) { Console.Write(" Empty request."); SendError(Client, 302, "The request wasn't heard. Please try again.", "\nLocation: "); return; }
-
-					RequestHeadersEnd = Request.IndexOf("\r\n\r\n");
-					if (RequestHeadersEnd == -1) RequestHeadersEnd = Request.Length;
-					RequestHeaders = Request.Substring(0, RequestHeadersEnd);
-					if (RequestHeadersEnd != Request.Length)
-						RequestBody = Request.Substring(RequestHeadersEnd + 4);
-					else
-					{ RequestBody = ""; Console.Write("[Short request]"); }
-					//may be caused by TCP packet non-gluing. Long HTTP requests, splitted to few packets, are read only on first packet. Bug!
-
-					/*Console.Write("-{0} of {1}-", Request.IndexOf("\r\n\r\n"), Request.Length);
-					Console.Write("-POST body={0}-", Request.Substring(Request.IndexOf("\r\n\r\n")).Trim("\r\n\r\n".ToCharArray()));*/
-				}
-				catch (System.IO.IOException ioe)
-				{
-					if (!ConfigFile.HideClientErrors)
-						Console.WriteLine("Can't read from client: " + ioe.ToString());
-					SendError(Client, 500);
-					return;
-				}
-
-
-				//find URL, method and headers
-				Match ReqMatch = Regex.Match(Request, @"^\w+\s+([^\s]+)[^\s]*\s+HTTP/.*|");
-				if (ReqMatch == Match.Empty)
-				{
-					//If the request seems to be invalid, raise HTTP 400 error.
-					Console.WriteLine(" Invalid request");
-					SendError(Client, 400, "Invalid request");
-					return;
-				}
-
-				string RequestMethod = "";
-				string[] Headers = RequestHeaders.Split('\n');
-				RequestMethod = Headers[0].Substring(0, Headers[0].IndexOf(" "));
-
-				WebHeaderCollection RequestHeaderCollection = new WebHeaderCollection();
-				foreach (string hdr in Headers)
-				{
-					if (hdr.Contains(": ")) //exclude method & URL
-					{
-						RequestHeaderCollection.Add(hdr.Replace("\r","")); //todo: add non-latin character urlencoding to prevent ArgumentException on cyrillic URLs
-					}
-					//if (hdr.Contains("ookie")) Console.WriteLine("Sent cookie: " + hdr);
-				}
+				ShouldRedirectInNETFW = false;
 
 				//check for login to proxy if need
 				if (ConfigFile.Authenticate != "")
 				{
-					if (RequestHeaderCollection["Proxy-Authorization"] == null || RequestHeaderCollection["Proxy-Authorization"] == "")
+					if (ClientRequest.Headers["Proxy-Authorization"] == null || ClientRequest.Headers["Proxy-Authorization"] == "")
 					{
-						SendError(Client, 407, "Hello! This Web 2.0-to-1.0 proxy server is private. Please enter your credentials.", "\n" + @"Proxy-Authenticate: Basic realm=""Log in to WebOne""");
+						Console.WriteLine("{0}\t Unauthorized client.", GetTime(BeginTime));
+						ClientResponse.AddHeader("Proxy-Authenticate", @"Basic realm=""Log in to WebOne""");
+						SendError(407, "Hello! This Web 2.0-to-1.0 proxy server is private. Please enter your credentials.");
 						return;
 					}
 					else
 					{
-						string auth = Encoding.Default.GetString(Convert.FromBase64String(RequestHeaderCollection["Proxy-Authorization"].Substring(6)));
+						string auth = Encoding.Default.GetString(Convert.FromBase64String(ClientRequest.Headers["Proxy-Authorization"].Substring(6)));
 						if (auth != ConfigFile.Authenticate)
 						{
-							SendError(Client, 407, "Your password is not correct. Please try again.", "\n" + @"Proxy-Authenticate: Basic realm=""Your WebOne credentials are incorrect""");
+							Console.WriteLine("{0}\t Incorrect login: '{1}'.", GetTime(BeginTime), auth);
+							ClientResponse.AddHeader("Proxy-Authenticate", @"Basic realm=""Your WebOne credentials are incorrect""");
+							SendError(407, "Your password is not correct. Please try again.");
 							return;
 						}
 					}
 				}
 
 				//fix "carousels"
-				string RefererUri = RequestHeaderCollection["Referer"];
-				RequestUri = ReqMatch.Groups[1].Value;
+				string RefererUri = ClientRequest.Headers["Referer"];
+				RequestURL = ClientRequest.Url;
 
 				//check for local or internal URL
-				if (RequestUri.StartsWith("/"))
+				if (RequestURL.Host == "localhost" || RequestURL.Host == Environment.MachineName || RequestURL.Host == "127.0.0.1")
 				{
-					if (RequestUri.StartsWith("/!"))
+					if (RequestURL.PathAndQuery.StartsWith("/!"))
 					{
 						//internal URLs
-						Console.Write("Internal: " + RequestUri);
-						switch (RequestUri.Substring(2, RequestUri.Length - 3))
+						Console.WriteLine("{0}\t Internal: {1} ", GetTime(BeginTime), RequestURL.PathAndQuery);
+						switch (RequestURL.AbsolutePath)
 						{
-							case "codepages":
+							case "/!codepages/":
 								string codepages = "The following code pages are available: <br>\n" +
 												   "<table><tr><td><b>Name</b></td><td><b>#</b></td><td><b>Description</b></td></tr>\n";
 								codepages += "<tr><td><b>AsIs</b></td><td>0</td><td>Leave page's code page as is</td></tr>\n";
@@ -160,69 +101,52 @@ namespace WebOne
 										codepages += "<tr><td><b>" + cp.Name + "</b></td><td>" + cp.CodePage + "</td><td>" + cp.DisplayName + "</td></tr>\n";
 								}
 								codepages += "</table><br>Use any of these. Underlined are for your locale.";
-								SendError(Client, 200, codepages);
+								SendError(200, codepages);
 								break;
 							default:
-								SendError(Client, 200, "Unknown internal URL: " + RequestUri.Substring(2, RequestUri.Length - 3));
+								SendError(200, "Unknown internal URL: " + RequestURL.PathAndQuery);
 								break;
 						}
 						return;
 					}
-					Console.Write("Local:");
-					LocalMode = true;
-					if (RequestUri.StartsWith("/http:") || RequestUri.StartsWith("/https:"))
-						RequestUri = RequestUri.Substring(1); //local mode: http://localhost:80/http://example.com/indexr.shtml
+					//local proxy mode: http://localhost/http://example.com/indexr.shtml
+					if (RequestURL.LocalPath.StartsWith("/http:") || RequestURL.AbsoluteUri.StartsWith("/https:"))
+					{
+						RequestURL = new Uri(RequestURL.LocalPath.Replace("/http:/", "http://"));
+						Console.WriteLine("{0}\t Local: {1}", GetTime(BeginTime), RequestURL);
+					}
 					else
 					{
-						//bad local mode, try to use last used host: http://localhost:80/favicon.ico
-						List<int> includes = new List<int>();
-
-						int i = LastURL.IndexOf("/", 0);
-						while (i > -1)
-						{
-							includes.Add(i);
-							i = LastURL.IndexOf("/", i + 1);
-						}
-						if (includes.Count < 3)
-						{
-							Console.Write("Bad direct URL ''", RequestUri);
-							SendError(Client, 400, "Bad direct URL: " + RequestUri + "\n<br>Expected something like " + LastURL);
-							return;
-						}
-						RequestUri = "http://" + LastURL.Substring(includes[0] + 2, includes[2] - includes[0] - 2) + RequestUri;
+						//dirty local mode, try to use last used host: http://localhost/favicon.ico
+						RequestURL = new Uri("http://" + new Uri(LastURL).Host + RequestURL.LocalPath);
+						Console.WriteLine("{0}\t Dirty local: {1}", GetTime(BeginTime), RequestURL);
 					}
-
 				}
 
 				//check for HTTP-to-FTP requests
 				//https://support.microsoft.com/en-us/help/166961/how-to-ftp-with-cern-based-proxy-using-wininet-api
-				string ProtocolName = "";
-				if(RequestUri.IndexOf(":")>1) ProtocolName = RequestUri.Substring(0, RequestUri.IndexOf(":"));
 				string[] BadProtocols = { "ftp", "gopher", "wais" };
-				if (Program.CheckString(ProtocolName, BadProtocols))
+				if (CheckString(RequestURL.Scheme, BadProtocols))
 				{
-					Console.Write(" " + RequestUri + " is a CERN Proxy request.");
-					SendError(Client, 101, "Cannot work with " + ProtocolName.ToUpper() + " protocol. Please connect directly bypassing the proxy.", "\nUpgrade: " + ProtocolName.ToUpper());
+					Console.WriteLine("{0}\t CERN Proxy request to {1} detected.", GetTime(BeginTime), RequestURL.Scheme.ToUpper());
+					SendError(101, "Cannot work with " + RequestURL.Scheme.ToUpper() + " protocol. Please connect directly bypassing the proxy.");//, "\nUpgrade: " + RequestURL.Scheme.ToUpper());
 					return;
 				}
 
 				//dirty workarounds for HTTP>HTTPS redirection bugs
-				if ((RequestUri == RefererUri || RequestUri == LastURL) && RequestUri != "" && RequestMethod != "POST" && RequestMethod != "CONNECT" && !Program.CheckString(RequestUri, ConfigFile.ForceHttps))
+				if ((RequestURL.AbsoluteUri == RefererUri || RequestURL.AbsoluteUri == LastURL) && RequestURL.AbsoluteUri != "" && ClientRequest.HttpMethod != "POST" && ClientRequest.HttpMethod != "CONNECT" && !Program.CheckString(RequestURL.AbsoluteUri, ConfigFile.ForceHttps))
 				{
-					Console.Write("Carousel");
-					if (!LastURL.StartsWith("https") && !RequestUri.StartsWith("https")) //if http is gone, try https
-						RequestUri = "https" + RequestUri.Substring(4);
+					Console.WriteLine("{0}\t Carousel detected.", GetTime(BeginTime));
+					if (!LastURL.StartsWith("https") && !RequestURL.AbsoluteUri.StartsWith("https")) //if http is gone, try https
+						RequestURL = new Uri("https" + RequestURL.AbsoluteUri.Substring(4));
 					if (LastURL.StartsWith("https")) //if can't use https, try again http
-						RequestUri = "http" + RequestUri.Substring(4);
+						RequestURL = new Uri("http" + RequestURL.AbsoluteUri.Substring(4));
 				}
-
-				Console.Write(" " + RequestUri + " ");
-				LastURL = RequestUri;
 
 				//check for too new frameworks & replace with older versions
 				foreach (string str in ConfigFile.FixableURLs)
 				{
-					if (Regex.Match(RequestUri, str).Success)
+					if (Regex.Match(RequestURL.AbsoluteUri, str).Success)
 					{
 						try
 						{
@@ -235,13 +159,15 @@ namespace WebOne
 							string InternalRedirect = "no";
 							if (ConfigFile.FixableUrlActions[str].ContainsKey("Internal")) InternalRedirect = ConfigFile.FixableUrlActions[str]["Internal"];
 
-							if (ValidMask == "" || !Regex.Match(RequestUri, ValidMask).Success/*!RequestUri.Contains(ConfigFile.FixableUrlActions[str]["ValidMask"])*/)
+							if (ValidMask == "" || !Regex.Match(RequestURL.AbsoluteUri, ValidMask).Success)
 							{
-								string NewURL = Redirect.Replace("%URL%", RequestUri).Replace("%UrlNoDomain%", RequestUri.Substring(RequestUri.IndexOf("/") + 2).Substring((RequestUri.Substring(RequestUri.IndexOf("/") + 2)).IndexOf("/") + 1));
+								string NewURL = Redirect.Replace("%URL%", RequestURL.AbsoluteUri).Replace("%UrlNoDomain%", RequestURL.AbsoluteUri.Substring(RequestURL.AbsoluteUri.IndexOf("/") + 2).Substring((RequestURL.AbsoluteUri.Substring(RequestURL.AbsoluteUri.IndexOf("/") + 2)).IndexOf("/") + 1));
+								//need to fix urlnodomain to extract using System.Uri instead of InStr/Mid$.
 
 								if (Redirect.Contains("%UrlNoPort%"))
 								{
-									Uri NewUri = new Uri(RequestUri);
+									
+									Uri NewUri = RequestURL;
 									var builder = new UriBuilder(NewUri);
 									builder.Port = -1;
 									NewUri = builder.Uri;
@@ -252,139 +178,126 @@ namespace WebOne
 									NewURL = NewUri.ToString();
 								}*/
 								
-								Console.Write("Fix to {1}", RequestUri, NewURL, ValidMask);
 								ShouldRedirectInNETFW = ConfigFile.ToBoolean(InternalRedirect);
 
 								if (!ShouldRedirectInNETFW)
 								{
-									SendError(Client, 302, "Брось каку!", "\nLocation: " + NewURL);
+									Console.WriteLine("{0}\t Fix to {1}", GetTime(BeginTime), NewURL);
+									ClientResponse.AddHeader("Location", NewURL);
+									SendError(302, "Брось каку!");
 									return;
 								}
 								else
 								{
-									RequestUri = NewURL;
-									Console.Write(" internally. ");
+									Console.WriteLine("{0}\t Fix to {1} internally", GetTime(BeginTime), NewURL);
+									RequestURL = new Uri(NewURL);
 								}
 							}
 						}
 						catch (Exception rex)
 						{
-							Console.Write("Cannot redirect! " + rex.Message);
-							SendError(Client, 200, rex.ToString().Replace("\n", "\n<br>"));
+							Console.WriteLine("{0}\t Cannot redirect! {1}", GetTime(BeginTime), rex.Message);
+							SendError(200, rex.ToString().Replace("\n", "\n<br>"));
 						}
 					}
 				}
 
-				//make reply
-				//SendError(Client, 200);
 
+				LastURL = RequestURL.AbsoluteUri;
+
+				//make reply
+				//SendError(200, "Okay, bro! Open " + RequestURL);
 				HTTPC https = new HTTPC();
 				bool StWrong = false; //break operation if something is wrong.
-				Console.Write("Try to " + RequestMethod.ToLower());
 
-				if (RequestUri.Contains("??")) { StWrong = true; SendError(Client, 400, "Too many questions."); }
-				if (RequestUri.Length == 0) { StWrong = true; SendError(Client, 400, "Empty URL."); }
-				if (RequestUri == "") return;
+				if (RequestURL.AbsoluteUri.Contains("??")) { StWrong = true; SendError(400, "Too many questions."); }
+				if (RequestURL.AbsoluteUri.Length == 0) { StWrong = true; SendError(400, "Empty URL."); }
+				if (RequestURL.AbsoluteUri == "") return;
 
 				try
 				{
 					int CL = 0;
-					if (RequestHeaderCollection["Content-Length"] != null) CL = Int32.Parse(RequestHeaderCollection["Content-Length"]);
+					if (ClientRequest.Headers["Content-Length"] != null) CL = Int32.Parse(ClientRequest.Headers["Content-Length"]);
 
 					//send HTTPS request to destination server
-					SendRequest(https, RequestMethod, RequestHeaderCollection, CL, Client);
+					WebHeaderCollection whc = new WebHeaderCollection();
+					foreach(string h in ClientRequest.Headers.Keys) {
+						whc.Add(h, ClientRequest.Headers[h]);
+					}
+					SendRequest(https, ClientRequest.HttpMethod, whc, CL);
 				}
 				catch (WebException wex)
 				{
 					if (wex.Response == null) ResponseCode = 502;
 					else ResponseCode = (int)(wex.Response as HttpWebResponse).StatusCode;
-					Console.Write("...WEB EXCEPTION=");
-					if (ResponseCode == 502) Console.Write(wex.Status); else Console.Write(ResponseCode + " " + (wex.Response as HttpWebResponse).StatusCode);
+					if (ResponseCode == 502) Console.WriteLine("{0}\t Cannot load this page: {1}.", GetTime(BeginTime), wex.Status);
+					else Console.WriteLine("{0}\t Web exception: {1} {2}.", GetTime(BeginTime), ResponseCode, (wex.Response as HttpWebResponse).StatusCode);
 
+					ContentType = "text/html";
 					string err = ": " + wex.Status.ToString();
-					/*if(wex.Response != null)
-					{
-						err = " because HTTP status code is not okay: " + (wex.Response as HttpWebResponse).StatusCode.ToString();
-
-						switch((wex.Response as HttpWebResponse).StatusCode) {
-							case HttpStatusCode.NotModified:
-								SendError(Client, 304, "Not modified, see in cache.");
-								StWrong = true;
-								break;
-							case HttpStatusCode.NotFound:
-								SendError(Client, 404, "The requested document is not found on server.");
-								StWrong = true;
-								break;
-							case HttpStatusCode.Forbidden:
-								SendError(Client, 403, "You're not welcome here.");
-								StWrong = true;
-								break;
-							default:
-								break;
-						}
-					}*/
 #if DEBUG
-					ResponseBody = "<html><body>Cannot load this page" + err + "<br><i>" + wex.ToString().Replace("\n", "<br>") + "</i><br>URL: " + RequestUri + Program.GetInfoString() + "</body></html>";
+					ResponseBody = "<html><body>Cannot load this page" + err + "<br><i>" + wex.ToString().Replace("\n", "<br>") + "</i><br>URL: " + RequestURL.AbsoluteUri + Program.GetInfoString() + "</body></html>";
 #else
-					ResponseBody = "<html><body>Cannot load this page" + err + " (<i>" + wex.Message + "</i>)<br>URL: " + RequestUri + Program.GetInfoString() + "</body></html>";
+					ResponseBody = "<html><body>Cannot load this page" + err + " (<i>" + wex.Message + "</i>)<br>URL: " + RequestURL.AbsoluteUri + Program.GetInfoString() + "</body></html>";
 #endif
 
 					//check if archived copy can be retreived instead
 					bool Archived = false;
 					if (ConfigFile.SearchInArchive)
-					if ((wex.Status == WebExceptionStatus.NameResolutionFailure) ||
-						(wex.Response != null && (wex.Response as HttpWebResponse).StatusCode == HttpStatusCode.NotFound))
-					{
-						try
+						if ((wex.Status == WebExceptionStatus.NameResolutionFailure) ||
+							(wex.Response != null && (wex.Response as HttpWebResponse).StatusCode == HttpStatusCode.NotFound))
 						{
-							Console.Write(". Look in Archive.org...");
-							HttpWebRequest ArchiveRequest = (HttpWebRequest)WebRequest.Create("https://archive.org/wayback/available?url=" + RequestUri);
-							ArchiveRequest.UserAgent = "WebOne/" + System.Reflection.Assembly.GetExecutingAssembly().GetName().Version;
-							ArchiveRequest.Method = "GET";
-							string ArchiveResponse = "";
-							MemoryStream ms = new MemoryStream();
-							ArchiveRequest.GetResponse().GetResponseStream().CopyTo(ms);
-							ArchiveResponse = Encoding.UTF8.GetString(ms.ToArray());
-							string ArchiveURL = "";
-
-							//parse archive.org json reply
-							if (ArchiveResponse.Contains(@"""available"": true"))
+							try
 							{
-								Console.Write("Available ");
-								Archived = true;
-								Match ArchiveMatch = Regex.Match(ArchiveResponse, @"""url"": ""http://web.archive.org/.*"","); ;
-								if (ArchiveMatch.Success)
+								Console.WriteLine("{0}\t Look in Archive.org...", GetTime(BeginTime));
+								HttpWebRequest ArchiveRequest = (HttpWebRequest)WebRequest.Create("https://archive.org/wayback/available?url=" + RequestURL.AbsoluteUri);
+								ArchiveRequest.UserAgent = "WebOne/" + System.Reflection.Assembly.GetExecutingAssembly().GetName().Version;
+								ArchiveRequest.Method = "GET";
+								string ArchiveResponse = "";
+								MemoryStream ms = new MemoryStream();
+								ArchiveRequest.GetResponse().GetResponseStream().CopyTo(ms);
+								ArchiveResponse = Encoding.UTF8.GetString(ms.ToArray());
+								string ArchiveURL = "";
+
+								//parse archive.org json reply
+								if (ArchiveResponse.Contains(@"""available"": true"))
 								{
-									ArchiveURL = ArchiveMatch.Value.Substring(8, ArchiveMatch.Value.Length - 10);
-									ResponseBody = "<html><body><h1>Server not found</h2>But an <a href=" + ArchiveURL + ">archived copy</a> is available! Redirecting to it...</body></html>";
-									SendError(Client, 302, ResponseBody, "\nLocation: " + ArchiveURL);
+									Archived = true;
+									Match ArchiveMatch = Regex.Match(ArchiveResponse, @"""url"": ""http://web.archive.org/.*"","); ;
+									if (ArchiveMatch.Success)
+									{
+										Console.WriteLine("{0}\t Available.", GetTime(BeginTime));
+										ArchiveURL = ArchiveMatch.Value.Substring(8, ArchiveMatch.Value.IndexOf(@""",") - 8);
+										ResponseBody = "<html><body><h1>Server not found</h2>But an <a href=" + ArchiveURL + ">archived copy</a> is available! Redirecting to it...</body></html>";
+										ClientResponse.AddHeader("Location", ArchiveURL);
+										SendError(302, ResponseBody);
+									}
+									else
+									{
+										Archived = false;
+										Console.WriteLine("{0}\t Available, but somewhere.", GetTime(BeginTime));
+									}
 								}
 								else
 								{
-									Archived = false;
-									Console.Write(" but somewhere");
-								}
-							}
-							else
-							{
-								Console.Write("No snapshots");
-								if (RequestUri.StartsWith("http://web.archive.org/web/") && ConfigFile.ShortenArchiveErrors)
-								{
+									Console.WriteLine("{0}\t No snapshots.", GetTime(BeginTime));
+									if (RequestURL.AbsoluteUri.StartsWith("http://web.archive.org/web/") && ConfigFile.ShortenArchiveErrors)
+									{
 										string ErrMsg =
 										"<p><b>The Wayback Machine has not archived that URL.</b></p>" +
-										"<p>This page is not available on the web because page does not exist<br>" + 
+										"<p>This page is not available on the web because page does not exist<br>" +
 										"Try to slightly change the URL.</p>" +
 										"<small><i>You see this message because ShortenArchiveErrors option is enabled.</i></small>";
-										SendError(Client, 404, ErrMsg);
+										SendError(404, ErrMsg);
 										Archived = true;
+									}
 								}
 							}
+							catch (Exception ArchiveException)
+							{
+								ResponseBody = String.Format("<html><body><b>Server not found and a Web Archive error occured.</b><br>{0}</body></html>", ArchiveException.Message.Replace("\n", "<br>"));
+							}
 						}
-						catch (Exception ArchiveException)
-						{
-							ResponseBody = String.Format("<html><body><b>Server not found and a Web Archive error occured.</b><br>{0}</body></html>", ArchiveException.Message.Replace("\n","<br>"));
-						}
-					}
 
 					//check if there are any response and the error isn't fatal
 					if (wex.Response != null)
@@ -394,77 +307,100 @@ namespace WebOne
 							string header = wex.Response.Headers.GetKey(i);
 							foreach (string value in wex.Response.Headers.GetValues(i))
 							{
-								if (!header.StartsWith("Connection") && !header.StartsWith("Transfer-Encoding"))
+								if (!header.StartsWith("Content-") &&
+								!header.StartsWith("Connection") &&
+								!header.StartsWith("Transfer-Encoding") &&
+								!header.StartsWith("Access-Control-Allow-Methods") &&
+								!header.StartsWith("Strict-Transport-Security") &&
+								!header.StartsWith("Content-Security-Policy") &&
+								!header.StartsWith("Upgrade-Insecure-Requests") &&
+								!(header.StartsWith("Vary") && value.Contains("Upgrade-Insecure-Requests")))
 								{
-									ResponseHeaders += (header + ": " + value.Replace("; secure", "").Replace("no-cache=\"set-cookie\"", "") + "\n");
-									if (RequestMethod == "OPTIONS" && header == "Allow") Console.Write("[Options allowed: {0}]", value);
+									ClientResponse.AddHeader(header, value.Replace("; secure", "").Replace("https://", "http://"));
 								}
 							}
 						}
 						ContentType = wex.Response.ContentType;
-						ResponseBody = new StreamReader(wex.Response.GetResponseStream()).ReadToEnd();
-						//ResponseBody = Encoding.UTF8.GetString(Encoding.UTF8.GetBytes(new StreamReader(wex.Response.GetResponseStream()).ReadToEnd()));
-						//todo: add correct returning of error pages through MakeOutput subprogram.
+						MakeOutput((HttpStatusCode)ResponseCode, wex.Response.GetResponseStream(),wex.Response.ContentType, wex.Response.ContentLength);
 					}
-
-					if (!Archived)
-					Console.WriteLine(".Failed[{0}].", ResponseCode);
+					#if DEBUG
+					else if (!Archived)
+						Console.WriteLine("{0}\t Failed: {1}.", GetTime(BeginTime), ResponseCode);
+					#endif
 				}
 				catch (UriFormatException)
 				{
 					StWrong = true;
-					SendError(Client, 400, "The URL <b>" + RequestUri + "</b> is not valid.");
+					SendError(400, "The URL <b>" + RequestURL.AbsoluteUri + "</b> is not valid.");
 				}
 				catch (Exception ex)
 				{
 					StWrong = true;
-					Console.WriteLine("============GURU MEDITATION:\n{0}\nOn URL '{1}', Method '{2}'. Returning 500.============", ex.ToString(), RequestUri, RequestMethod);
-					SendError(Client, 500, "Guru meditaion at URL " + RequestUri + ":<br><b>" + ex.Message + "</b><br><i>" + ex.StackTrace.Replace("\n", "\n<br>") + "</i>");
+					Console.WriteLine("{0}\t ============GURU MEDITATION:\n{1}\nOn URL '{2}', Method '{3}'. Returning 500.============", GetTime(BeginTime), ex.ToString(), RequestURL.AbsoluteUri, ClientRequest.HttpMethod);
+					SendError(500, "Guru meditaion at URL " + RequestURL.AbsoluteUri + ":<br><b>" + ex.Message + "</b><br><i>" + ex.StackTrace.Replace("\n", "\n<br>") + "</i>");
 				}
 
+				//try to return...
 				try
 				{
-					//try to return...
 					if (!StWrong)
 					{
-						ResponseHeaders += "Via: HTTP/1.0 WebOne/" + System.Reflection.Assembly.GetExecutingAssembly().GetName().Version + "\n";
-						ResponseHeaders += "Connection: close\n";
-						//ResponseHeaders += "Warning: 214 WebOne/" + System.Reflection.Assembly.GetExecutingAssembly().GetName().Version + @" ""Patched for old browser""" + "\n";
+						ClientResponse.ProtocolVersion = new Version(1, 0);
+						ClientResponse.StatusCode = ResponseCode;
+						ClientResponse.AddHeader("Via", "HTTP/1.0 WebOne/" + System.Reflection.Assembly.GetExecutingAssembly().GetName().Version);
 
-						if (Program.CheckString(ContentType, ConfigFile.TextTypes) || ContentType == "")
-							ResponseHeaders = "HTTP/1.0 " + ResponseCode + "\n" + ResponseHeaders + "Content-Type: " + ContentType + "\nContent-Length: " + ResponseBody.Length;
+
+						if (CheckString(ContentType, ConfigFile.TextTypes) || ContentType == "")
+						{
+							ClientResponse.AddHeader("Content-Type", ContentType);
+						}
 						else
-							ResponseHeaders = "HTTP/1.0 " + ResponseCode + "\n" + ResponseHeaders + "Content-Type: " + ContentType + "\nContent-Length: " + response.ContentLength;
+						{
+							ClientResponse.AddHeader("Content-Type", ContentType);
+						}
 
-						byte[] RespBuffer = (ConfigFile.OutputEncoding ?? Encoding.Default).GetBytes(ResponseHeaders + "\n\n");
 						if (TransitStream == null)
 						{
-							RespBuffer = RespBuffer.Concat(ResponseBuffer ?? (ConfigFile.OutputEncoding ?? Encoding.Default).GetBytes(ResponseBody)).ToArray();
-							ClientStream.Write(RespBuffer, 0, RespBuffer.Length);
+							byte[] RespBuffer;
+							RespBuffer = (ConfigFile.OutputEncoding ?? Encoding.Default).GetBytes(ResponseBody).ToArray();
+
+							if(ClientResponse.ContentLength64 > 300*1024) Console.WriteLine("{0}\t Sending binary.", GetTime(BeginTime));
+							ClientResponse.OutputStream.Write(RespBuffer, 0, RespBuffer.Length);
 						}
 						else
 						{
-							ClientStream.Write(RespBuffer, 0, RespBuffer.Length);
-							TransitStream.CopyTo(ClientStream);
+							TransitStream.CopyTo(ClientResponse.OutputStream);
 						}
-						Client.Close();
+						ClientResponse.OutputStream.Close();
+						#if DEBUG
+						Console.WriteLine("{0}\t Document sent.", GetTime(BeginTime));
+						#endif
 					}
+					#if DEBUG
+					else Console.WriteLine("{0}\t Abnormal return (something was wrong).", GetTime(BeginTime));
+					#endif
 				}
 				catch (Exception ex)
 				{
-					if(!ConfigFile.HideClientErrors)
-					Console.WriteLine("Cannot return reply to the client. " + ex.Message + ex.StackTrace);
+					if (!ConfigFile.HideClientErrors)
+						#if DEBUG
+						Console.WriteLine("{0}\t<Can't return reply. " + ex.Message + ex.StackTrace, GetTime(BeginTime));
+						#else
+						Console.WriteLine("{0}\t<Can't return reply. " + ex.Message, GetTime(BeginTime));
+						#endif
 				}
-
 			}
-			catch(Exception ex)
+			catch(Exception E)
 			{
-				Console.WriteLine("A error has been catched: {0}\nPlease report to author.", ex.ToString());
-				SendError(Client, 500, "WTF?! " + ex.ToString().Replace("\n", "\n<BR>"));
-				return;
+				string time = GetTime(BeginTime);
+				Console.WriteLine("{0}\t A error has been catched: {1}\n{0}\t Please report to author.", time, E.ToString().Replace("\n", "\n{0}\t "));
+				SendError(500, "WTF?! " + E.ToString().Replace("\n", "\n<BR>"));
 			}
-			Console.WriteLine("\nThe client is served.");
+			#if DEBUG
+			Console.WriteLine("{0}\t End process.", GetTime(BeginTime));
+			#endif
 		}
+
 
 		/// <summary>
 		/// Send a HTTPS request and put the response to shared variable "response"
@@ -473,45 +409,38 @@ namespace WebOne
 		/// <param name="RequestMethod">Request method</param>
 		/// <param name="RequestHeaderCollection">Request headers</param>
 		/// <param name="Content_Length">Request content length</param>
-		/// <param name="Client">TCP client</param>
 		/// <returns>Response status code (and the Response in shared variable)</returns>
-		private void SendRequest(HTTPC https, string RequestMethod, WebHeaderCollection RequestHeaderCollection, int Content_Length, TcpClient Client)
+		private void SendRequest(HTTPC https, string RequestMethod, WebHeaderCollection RequestHeaderCollection, int Content_Length)
 		{
-			//moved from constructor as is on july 19, 2019
-			//probably needs to be cleaned up and rewritten
-			bool AllowAutoRedirect = Program.CheckString(RequestUri, ConfigFile.InternalRedirectOn);
+			bool AllowAutoRedirect = CheckString(RequestURL.AbsoluteUri, ConfigFile.InternalRedirectOn);
 			if (!AllowAutoRedirect) AllowAutoRedirect = ShouldRedirectInNETFW;
+			string RequestBody;
+
 			switch (RequestMethod)
 			{
-				case "GET":
-					//try to get...
-					response = https.GET(RequestUri, new CookieContainer(), RequestHeaderCollection, "GET", AllowAutoRedirect);
-					MakeOutput(response);
-					break;
-				case "POST":
-					//try to post...
-					response = https.POST(RequestUri, new CookieContainer(), RequestBody, RequestHeaderCollection, "POST", AllowAutoRedirect);
-					MakeOutput(response);
-					break;
 				case "CONNECT":
 					string ProtocolReplacerJS = "<script>if (window.location.protocol != 'http:') { setTimeout(function(){window.location.protocol = 'http:'; window.location.reload();}, 1000); }</script>";
-					SendError(Client, 405, "The proxy does not know the " + RequestMethod + " method.<BR>Please use HTTP, not HTTPS.<BR>HSTS must be disabled." + ProtocolReplacerJS);
-					Console.WriteLine(" Wrong method.");
+					SendError(405, "The proxy does not know the " + RequestMethod + " method.<BR>Please use HTTP, not HTTPS.<BR>HSTS must be disabled." + ProtocolReplacerJS);
+					Console.WriteLine("{0}\t Wrong method.", GetTime(BeginTime));
 					return;
 				default:
 					if (Content_Length == 0)
 					{
-						//try to download (HEAD, WebDAV download, etc)
-						response = https.GET(RequestUri, new CookieContainer(), RequestHeaderCollection, RequestMethod, AllowAutoRedirect);
-						MakeOutput(response);
+						//try to download (GET, HEAD, WebDAV download, etc)
+						Console.WriteLine("{0}\t>Downloading content...", GetTime(BeginTime));
+						response = https.GET(RequestURL.AbsoluteUri, new CookieContainer(), RequestHeaderCollection, RequestMethod, AllowAutoRedirect);
+						MakeOutput(response.StatusCode, response.Stream, response.ContentType, response.ContentLength);
 						break;
 					}
 					else
 					{
-						//try to upload (PUT, WebDAV, etc)
-						Console.Write(" CL={0}K", Convert.ToInt32((RequestHeaderCollection["Content-Length"])) / 1024);
-						response = https.POST(RequestUri, new CookieContainer(), RequestBody, RequestHeaderCollection, RequestMethod, AllowAutoRedirect);
-						MakeOutput(response);
+						//try to upload (POST, PUT, WebDAV, etc)
+						Console.WriteLine("{0}\t>Reading input stream...", GetTime(BeginTime));
+						StreamReader body_sr = new StreamReader(ClientRequest.InputStream);
+						RequestBody = body_sr.ReadToEnd();
+						Console.WriteLine("{0}\t>Uploading {1}K...", GetTime(BeginTime), Convert.ToInt32((RequestHeaderCollection["Content-Length"])) / 1024);
+						response = https.POST(RequestURL.AbsoluteUri, new CookieContainer(), RequestBody, RequestHeaderCollection, RequestMethod, AllowAutoRedirect);
+						MakeOutput(response.StatusCode, response.Stream, response.ContentType, response.ContentLength);
 						break;
 					}
 			}
@@ -521,20 +450,20 @@ namespace WebOne
 			//check for security upgrade
 			if (ResponseCode == 301 || ResponseCode == 302 || ResponseCode == 308)
 			{
-				if (RequestUri == (response.Headers["Location"] ?? "nowhere").Replace("https://", "http://")
-					&& !Program.CheckString(RequestUri, ConfigFile.InternalRedirectOn))
+				if (RequestURL.AbsoluteUri == (response.Headers["Location"] ?? "nowhere").Replace("https://", "http://")
+					&& !CheckString(RequestURL.AbsoluteUri, ConfigFile.InternalRedirectOn))
 				{
-					Console.Write("\n> Reload secure");
-					RequestUri = RequestUri.Replace("http://", "https://");
+					Console.WriteLine("{0}\t>Reload secure...", GetTime(BeginTime));
+
+					RequestURL = new Uri(RequestURL.AbsoluteUri.Replace("http://", "https://"));
 					TransitStream = null;
-					SendRequest(https, RequestMethod, RequestHeaderCollection, Content_Length, Client);
-					//ResponseHeaders += "X-Redirected: Make-Https\n";
+					SendRequest(https, RequestMethod, RequestHeaderCollection, Content_Length);
 
 					//add to ForceHttp list
 					List<string> ForceHttpsList = ConfigFile.ForceHttps.ToList<string>();
-					string SecureHost = new Uri(RequestUri).Host;
-					if(!ForceHttpsList.Contains(SecureHost))
-					ForceHttpsList.Add(SecureHost);
+					string SecureHost = RequestURL.Host;
+					if (!ForceHttpsList.Contains(SecureHost))
+						ForceHttpsList.Add(SecureHost);
 					ConfigFile.ForceHttps = ForceHttpsList.ToArray();
 
 					return;
@@ -555,77 +484,22 @@ namespace WebOne
 					!header.StartsWith("Upgrade-Insecure-Requests") &&
 					!(header.StartsWith("Vary") && value.Contains("Upgrade-Insecure-Requests")))
 					{
-						ResponseHeaders += (header + ": " + value.Replace("; secure", "") + "\n").Replace("https://", "http://");
-						//Console.WriteLine(header + ": " + value.Replace("; secure", "").Replace("no-cache=\"set-cookie\"", ""));
-						//if (header.Contains("ookie")) Console.WriteLine("Got cookie: " + value);
+						ClientResponse.AddHeader(header, value.Replace("; secure", "").Replace("https://", "http://"));
 					}
 				}
 			}
-
-			/*if (RequestMethod=="OPTIONS")
-			{
-				ResponseHeaders += "Access-Control-Allow-Methods: POST, GET, OPTIONS\n";
-			}*/
 		}
 
-		/// <summary>
-		/// Make reply's body as byte array
-		/// </summary>
-		/// <param name="response">The HTTP Response</param>
-		/// <returns>ResponseBuffer+ResponseBody for texts or TransitStream for binaries</returns>
-		private void MakeOutput(HttpResponse response) {
-			Console.Write("...");
-			Console.Write(response.StatusCode);
-			ContentType = response.ContentType;
-			Console.Write("...Body {0}K of {1}.", response.ContentLength / 1024, ContentType);
-
-			if (Program.CheckString(ContentType, ConfigFile.TextTypes))
-			{
-				//если сервер вернул текст, сделать правки, иначе прогнать как есть дальше
-				Console.Write("[Text]");
-
-				if (ConfigFile.OutputEncoding == null)
-				{
-					//if don't touch codepage (OutputEncoding=AsIs)
-					ResponseBody = Encoding.Default.GetString(response.RawContent);
-					ResponseBody = ProcessBody(ResponseBody);
-					ResponseBuffer = Encoding.Default.GetBytes(ResponseBody);
-					Console.WriteLine(".");
-					return;
-				}
-
-				bool ForceUTF8 = ContentType.ToLower().Contains("utf-8");
-				if(!ForceUTF8 && response.RawContent.Length > 0) ForceUTF8 = response.RawContent[0] == Encoding.UTF8.GetPreamble()[0];
-				foreach (string utf8url in ConfigFile.ForceUtf8) { if (Regex.IsMatch(RequestUri, utf8url)) ForceUTF8 = true; }
-				//todo: add fix for "invalid range in character class" at www.yandex.ru with Firefox 3.6 if OutputEncoding!=AsIs
-
-				if (ForceUTF8) ResponseBody = Encoding.UTF8.GetString(response.RawContent);
-				else ResponseBody = Encoding.Default.GetString(response.RawContent);
-				
-				if (Regex.IsMatch(ResponseBody, @"<meta.*UTF-8.*>", RegexOptions.IgnoreCase)) { ResponseBody = Encoding.UTF8.GetString(response.RawContent); }
-				//ResponseBuffer = ConfigFile.OutputEncoding.GetBytes(ResponseBody);
-
-				if (ContentType.ToLower().Contains("utf-8")) ContentType = ContentType.Substring(0, ContentType.IndexOf(';'));
-				ResponseBody = ProcessBody(ResponseBody);
-				ResponseBuffer = ConfigFile.OutputEncoding.GetBytes(ResponseBody);
-			}
-			else
-			{
-				Console.Write("[Binary]");
-				TransitStream = response.Stream;
-			}
-			Console.Write(".");
-			return;
-		}
 
 		/// <summary>
 		/// Process the reply's body and fix too modern stuff
 		/// </summary>
 		/// <param name="Body">The original body</param>
 		/// <returns>The fixed body, compatible with old browsers</returns>
-		private string ProcessBody(string Body) {
+		private string ProcessBody(string Body)
+		{
 			Body = Body.Replace("https", "http");
-			if(LocalMode) Body = Body.Replace("http://", "http://" + Environment.MachineName + "/http://");//replace with real hostname
+			//if (LocalMode) Body = Body.Replace("http://", "http://" + Environment.MachineName + "/http://");//replace with real hostname
 			if (ConfigFile.OutputEncoding != null)
 			{
 				Body = Body.Replace("harset=\"utf-8\"", "harset=\"" + ConfigFile.OutputEncoding.WebName + "\"");
@@ -647,18 +521,18 @@ namespace WebOne
 			foreach (string mask in ConfigFile.ContentPatches)
 			{
 				string IfURL = ".*";
-				if(ConfigFile.ContentPatchActions[mask].ContainsKey("IfURL"))
+				if (ConfigFile.ContentPatchActions[mask].ContainsKey("IfURL"))
 					IfURL = ConfigFile.ContentPatchActions[mask]["IfURL"];
-				
+
 				string IfType = ".*";
-				if(ConfigFile.ContentPatchActions[mask].ContainsKey("IfType"))
+				if (ConfigFile.ContentPatchActions[mask].ContainsKey("IfType"))
 					IfType = ConfigFile.ContentPatchActions[mask]["IfType"];
 
 				string Replace = "<!--Removed by WebOne:$&End-->";
 				if (ConfigFile.ContentPatchActions[mask].ContainsKey("Replace"))
 					Replace = ConfigFile.ContentPatchActions[mask]["Replace"];
 
-				if(Regex.IsMatch(RequestUri,IfURL) && Regex.IsMatch(ContentType, IfType))
+				if (Regex.IsMatch(RequestURL.AbsoluteUri, IfURL) && Regex.IsMatch(ContentType, IfType))
 				{
 					try
 					{
@@ -667,53 +541,96 @@ namespace WebOne
 					}
 					catch (Exception rex)
 					{
-						Console.Write(" Cannot make edit: '" + rex.Message + "'!");
+						Console.WriteLine("{0}\t Cannot make edit: {1}!", GetTime(BeginTime), rex.Message);
 					}
 				}
 			}
 
-			if (patched > 0) Console.Write(" {0} patches applied", patched);
+			if (patched > 0) Console.WriteLine("{0}\t {1} patch(-es) applied...", GetTime(BeginTime), patched);
 			return Body;
 		}
+
+
+		/// <summary>
+		/// Prepare response body for tranfer to client
+		/// </summary>
+		/// <param name="StatusCode">HTTP Status code</param>
+		/// <param name="ResponseStream">Stream of response body</param>
+		/// <param name="ContentType">HTTP Content-Type</param>
+		/// <param name="ContentLength">HTTP Content-Lenght</param>
+		/// <returns>ResponseBuffer+ResponseBody for texts or TransitStream for binaries</returns>
+		private void MakeOutput(HttpStatusCode StatusCode, Stream ResponseStream, string ContentType, long ContentLength)
+		{
+			if (Program.CheckString(ContentType, ConfigFile.TextTypes))
+			{
+				//if server returns text, make edits
+				Console.WriteLine("{0}\t {1} {2}. Body {3}K of {4} [Text].", GetTime(BeginTime), (int)StatusCode, StatusCode, ContentLength / 1024, ContentType);
+				byte[] RawContent = null;
+				RawContent = ReadAllBytes(ResponseStream);
+
+				if (ConfigFile.OutputEncoding == null)
+				{
+					//if don't touch codepage (OutputEncoding=AsIs)
+					ResponseBody = Encoding.Default.GetString(RawContent);
+					ResponseBody = ProcessBody(ResponseBody);
+					Console.WriteLine("{0}\t Body maked.", GetTime(BeginTime));
+					return;
+				}
+
+				bool ForceUTF8 = ContentType.ToLower().Contains("utf-8");
+				if (!ForceUTF8 && RawContent.Length > 0) ForceUTF8 = RawContent[0] == Encoding.UTF8.GetPreamble()[0];
+				foreach (string utf8url in ConfigFile.ForceUtf8) { if (Regex.IsMatch(RequestURL.AbsoluteUri, utf8url)) ForceUTF8 = true; }
+				//todo: add fix for "invalid range in character class" at www.yandex.ru with Firefox 3.6 if OutputEncoding!=AsIs
+
+				if (ForceUTF8) ResponseBody = Encoding.UTF8.GetString(RawContent);
+				else ResponseBody = Encoding.Default.GetString(RawContent);
+
+				if (Regex.IsMatch(ResponseBody, @"<meta.*UTF-8.*>", RegexOptions.IgnoreCase)) { ResponseBody = Encoding.UTF8.GetString(RawContent); }
+
+				if (ContentType.ToLower().Contains("utf-8")) ContentType = ContentType.Substring(0, ContentType.IndexOf(';'));
+				ResponseBody = ProcessBody(ResponseBody);
+				this.ContentType = ContentType;
+			}
+			else
+			{
+				Console.WriteLine("{0}\t {1} {2}. Body {3}K of {4} [Binary].", GetTime(BeginTime), (int)StatusCode, StatusCode, response.ContentLength / 1024, ContentType);
+				TransitStream = ResponseStream;
+				this.ContentType = ContentType;
+			}
+			#if DEBUG
+			Console.WriteLine("{0}\t Body maked.", GetTime(BeginTime));
+			#endif
+			return;
+		}
+
 
 		/// <summary>
 		/// Send a HTTP error to client
 		/// </summary>
-		/// <param name="Client">TcpListener client</param>
-		/// <param name="Code">Error code number</param>
-		/// <param name="Text">Error description for user</param>
-		/// <param name="ExtraHeaders">Additional headers (like "Location: filename.ext" for 302 or "Refresh: 10; filename.ext" for 200)</param>
-		private void SendError(TcpClient Client, int Code, string Text = "", string ExtraHeaders = "")
+		/// <param name="Code">HTTP Status code</param>
+		/// <param name="Text">Text of message</param>
+		private void SendError(int Code, string Text = "")
 		{
-			Console.Write("["+Code+"]\n");
-			Text += Program.GetInfoString();
+			Console.WriteLine("{0}\t<Return code {1}.", GetTime(BeginTime), Code);
+			Text += GetInfoString();
 			string CodeStr = Code.ToString() + " " + ((HttpStatusCode)Code).ToString();
-			string Refresh = "<META HTTP-EQUIV=REFRESH CONTENT=0>";
-			if (Code != 302 || ExtraHeaders.StartsWith("Refresh:") || ExtraHeaders.StartsWith("Location:")) Refresh = "";
-			string Html = "<html>" + Refresh + "<body><h1>" + CodeStr + "</h1>"+Text+"</body></html>";
-			string Str = "HTTP/1.0 " + CodeStr + "\nContent-type: text/html\nContent-Length:" + Html.Length.ToString() + ExtraHeaders + "\n\n" + Html;
-			byte[] Buffer = Encoding.Default.GetBytes(Str);
+			string Refresh = "";
+			if (ClientResponse.Headers["Refresh"] != null) Refresh = "<META HTTP-EQUIV=REFRESH CONTENT="+ ClientResponse.Headers["Refresh"] +">";
+			string Html = "<html>" + Refresh + "<body><h1>" + CodeStr + "</h1>" + Text + "</body></html>";
+			byte[] Buffer = Encoding.Default.GetBytes(Html);
 			try
 			{
-				Client.GetStream().Write(Buffer, 0, Buffer.Length);
-				Client.Close();
-			}
-			catch {
-				Console.WriteLine("Cannot return HTTP error " + Code);
-			}
-		}
+				ClientResponse.StatusCode = Code;
+				ClientResponse.ProtocolVersion = new Version(1, 0);
 
-		/// <summary>
-		/// Read a Stream as byte array
-		/// </summary>
-		/// <param name="input">The stream to be read</param>
-		/// <returns></returns>
-		private byte[] ReadFully(Stream input)
-		{
-			using (MemoryStream ms = new MemoryStream())
+				ClientResponse.ContentType = "text/html";
+				ClientResponse.ContentLength64 = Html.Length;
+				ClientResponse.OutputStream.Write(Buffer, 0, Buffer.Length);
+				ClientResponse.OutputStream.Close();
+			}
+			catch
 			{
-				input.CopyTo(ms);
-				return ms.ToArray();
+				Console.WriteLine("{0}\t<!Cannot return code {1}.", GetTime(BeginTime), Code);
 			}
 		}
 	}
