+++ WebOne.csproj
@@ -52,11 +52,9 @@
     <Compile Include="ConfigFile.cs" />
     <Compile Include="HTTPC.cs" />
     <Compile Include="HTTPServer.cs" />
-    <Compile Include="NewHTTPServer.cs" />
-    <Compile Include="NewTransit.cs" />
+    <Compile Include="Transit.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Transit.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
